// src/pages/types/category.type.tsx
export type Category = {
  idCategory: number;
  title: string;
  subtitle: string;
  description: string;
  subDescription: string;
  subImage: string;
  parentId: number;
  hasChild: string;
  icon: string;
  image: string;
  exibitionType: string;
  groupName: string;
  viewOrder: number;
  requestPriority: number;
  categoryType: string;
  subImageLang: string;
};